pyproject.toml
[tool.poetry]
name = "grabcut-annotation-tool"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.5.5"
Pillow = "^9.0.0"
PySimpleGUI = "^4.56.0"
click = "^8.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
